% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{howManyCells}
\alias{howManyCells}
\title{Count the number of cells}
\usage{
howManyCells(
  seur,
  features,
  group.by = NULL,
  assay = "RNA",
  slot = "data",
  min_exp = 1,
  ident
)
}
\arguments{
\item{seur}{Seurat object to use.}

\item{features}{Features to use.}

\item{min_exp}{}
}
\value{

}
\description{
Count the number of cells
}
