% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_plot.R
\name{plotCumulativeVarExplained}
\alias{plotCumulativeVarExplained}
\title{Plot the cumulative variance explained}
\usage{
plotCumulativeVarExplained(seur, npcs = 50)
}
\arguments{
\item{seur}{The Seurat object from which the PCA data will be taken}

\item{npcs}{The number of PCs to plot}
}
\value{
A ggplot object
}
\description{
This function first verify that the variance explained has been computed,
and if not computes it. It then computes the cumulative sum of the variances,
and then creates a plot of the cumulative variance explained for each PC.
}
