% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_plot.R
\name{plotVarianceExplained}
\alias{plotVarianceExplained}
\title{Plot the variance explained}
\usage{
plotVarianceExplained(seur, npcs = 50)
}
\arguments{
\item{seur}{The Seurat object from which the PCA data will be taken}

\item{npcs}{The number of PCs to plot}
}
\value{
A ggplot object
}
\description{
This function first verify that the variance explained has been computed,
and if not computes it. It then creates a plot of the variance explained
for each PC.
}
