% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setCols}
\alias{setCols}
\title{Set colors}
\usage{
setCols(cols, group.by)
}
\arguments{
\item{cols}{NULL or a vector of colors (named or not)}

\item{group.by}{column with duplicated names of colors}
}
\value{
colors named
}
\description{
Set colors
}
