% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_plot.R
\name{theme_white_bg}
\alias{theme_white_bg}
\title{ggplot2 theme with white background and axis lines}
\usage{
theme_white_bg()
}
\value{
ggplot2 theme
}
\description{
ggplot2 theme with white background and axis lines
}
