% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WhichCells2}
\alias{WhichCells2}
\title{Count the number of cells}
\usage{
WhichCells2(
  seur,
  features,
  assay = "RNA",
  slot = "data",
  value_to_compare = 0,
  comparison_type = "superior",
  verbose = TRUE
)
}
\arguments{
\item{seur}{Seurat object to use.}

\item{features}{Features to use.}

\item{assay}{Assay to use if features are genes. Default to "RNA".}

\item{slot}{Slot to use if features are genes. Default to "data".}

\item{value_to_compare}{Which value (or values) to compare the features to.
Must be either a single element or a list of the same length as features.
Default to 0.}

\item{comparison_type}{one of "inferior" or 1, "inferior or equal" or 2,
"equal" or 3, "superior or equal" or 4, "superior" or 5.
Can be a list of them, the same length as features. Default to "superior".}

\item{verbose}{Whether or not to print the summary dataframe. Default to TRUE.}
}
\value{
The number of cells respecting all the conditions.
}
\description{
Count the number of cells respecting all the conditions, and print a recap with the number of cells respecting each condition.
}
