% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_plot.R
\name{plotGenesRank}
\alias{plotGenesRank}
\title{Plot the genes ranked by expression}
\usage{
plotGenesRank(
  seur,
  genes_highlight,
  group.by = "orig.ident",
  assay = "RNA",
  slot = "data",
  colmap = FALSE,
  nudge_x = "middle",
  ncol = 2
)
}
\arguments{
\item{seur}{Seurat object to use.}

\item{genes_highlight}{List of genes to highlight.
If some genes are not found in the dataset, they will just not be plotted.}

\item{group.by}{Which ident to group the data by. Default to "orig.ident".}

\item{assay}{Assay to use. Default to "RNA"}

\item{slot}{slot to take the data from. Using data will assume that the data
are log1p normalized, and thus will exp1p them before getting the average expression.
Default to "data"}

\item{colmap}{colmap to use for each ident. Default to hue_pal()}

\item{nudge_x}{either a number or "middle". Will nudge the gene name
on the right if positive and left if negative.
Middle will try to align the gene names in the middle. Default to 2}

\item{ncol}{number of column to plot, only useful if more than 1 ident.
Default to 2}
}
\value{
a ggplot object if only one ident, or a gridExtra object if more than one ident.
}
\description{
Genes will be ranked by expression, and then plotted, with x axis being the
rank and y axis the average expression in the cells, for each ident in group.by.
Some genes will then be highlighted.
}
